requirements.txt
discord
python-dotenv
SQLAlchemy
quart
psycopg2
asyncpg